refactored: refactored the AgendaModal file, implementing the atomic design pattern
